Client login

+++ server/routes/groups_routes.js
@@ -1,6 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const {db} = require('../services/db');
+const auth = require("../middleware/auth");
+const e = require('express');
 
   // Helper function to generate random alphanumeric code
   function generateCode(length) {
@@ -30,9 +32,9 @@ const {db} = require('../services/db');
   }
   
 
-router.get('/admin',async (req, res) => {
+router.get('/admin', auth, async (req, res) => {
   const SelectQuery = " SELECT * FROM  Carbon_Footprint_Groups WHERE owner_ID = ?";
-  db.query(SelectQuery, [req.query.user_ID], (err, result) => {
+  db.query(SelectQuery, [req.user.user_id], (err, result) => {
     //count the members of each group and add it to the result
     result.forEach(async (group, index) => {
         result[index].memberCount = await getMemberCount(group.group_ID);
@@ -46,9 +48,9 @@ router.get('/admin',async (req, res) => {
 })
 
   
-router.get('/member', async(req, res) => {
+router.get('/member', auth, async(req, res) => {
     const SelectQuery = " SELECT * FROM  Groupmemberships WHERE user_ID = ?";
-    db.query(SelectQuery, [req.query.user_ID], (err, result) => {
+    db.query(SelectQuery, [req.user.user_id], (err, result) => {
       // if result is empty, send empty array
       if (result.length === 0) {
         res.send([])
@@ -69,14 +71,14 @@ router.get('/member', async(req, res) => {
     })
   })
   
-router.post('/add_user', (req, res) => {
+router.post('/add_user', auth,(req, res) => {
     const SelectQuery = " SELECT group_ID FROM  Carbon_Footprint_Groups WHERE groupcode = ?";
-    db.query(SelectQuery, [req.query.groupcode], (err, result) => {
+    db.query(SelectQuery, [req.body.groupcode], (err, result) => {
       if (result.length === 0) {
         res.status(404).send('Group not found');
       } else {
         const InsertQuery = " INSERT INTO Groupmemberships (group_ID, user_ID) VALUES (?, ?)";
-        db.query(InsertQuery, [result[0].group_ID, req.query.user_ID], (err, result) => {
+        db.query(InsertQuery, [result[0].group_ID, req.user.user_id], (err, result) => {
           if(err) {
             if(err.code === 'ER_DUP_ENTRY') {
               console.log(err)
@@ -94,9 +96,10 @@ router.post('/add_user', (req, res) => {
     })
   })
   
-router.get('/get/:groupcode',async  (req, res) => {
+router.get('/get', async  (req, res) => {
+
     const SelectQuery = " SELECT * FROM  Carbon_Footprint_Groups WHERE groupcode = ?";
-    db.query(SelectQuery, [req.params.groupcode], async (err, result) => {
+    db.query(SelectQuery, [req.query.groupcode], async (err, result) => {
       if (result.length === 0) {
         res.status(404).send('Group not found');
   
@@ -106,15 +109,15 @@ router.get('/get/:groupcode',async  (req, res) => {
         result[0].memberCount = await getMemberCount(result[0].group_ID);
         delete result[0].user_ID;
         delete result[0].owner_ID;
-        delete result[0].groupcode;
         delete result[0].group_ID;
+        result[0].groupcode;
         res.send(result[0])
             
       }
     })
   })
   
-router.post('/create', async (req, res) => {
+router.post('/create', auth, async (req, res) => {
       let groupcode = '';
       let isUnique = false;
       while (!isUnique) {
@@ -128,35 +131,40 @@ router.post('/create', async (req, res) => {
         }
       }
       const InsertQuery = " INSERT INTO Carbon_Footprint_Groups (groupname, groupcode, owner_ID) VALUES (?, ?, ?)";
-      db.query(InsertQuery, [req.query.groupname, groupcode, req.query.user_ID], (err, result) => {
+      db.query(InsertQuery, [req.body.groupname, groupcode, req.user.user_id], (err, result) => {
         if(err) {
           console.log(err)
           res.status(500).send('Something went wrong')
         } else {
           const InsertQuery = " INSERT INTO Groupmemberships (group_ID, user_ID) VALUES (?, ?)";
-          db.query(InsertQuery, [result.insertId, req.query.user_ID], (err, result2) => {
+          db.query(InsertQuery, [result.insertId, req.user.user_id], (err, result2) => {
           res.status(201).json({ groupcode });
         }
         )
       }
       })
     })
-  
-router.delete('/delete/:groupcode', (req, res) => {
-    const SelectQuery = " SELECT group_ID FROM  Carbon_Footprint_Groups WHERE groupcode = ?";
-    db.query(SelectQuery, [req.params.groupcode], (err, result) => {
+
+router.delete('/delete',auth,  (req, res) => {
+    const SelectQuery = " SELECT group_ID, owner_ID FROM  Carbon_Footprint_Groups WHERE groupcode = ?";
+    db.query(SelectQuery, [req.body.groupcode], (err, result) => {
       if (result.length === 0) {
         res.status(404).send({ error : 'Group not found'});
       } else {
-        const DeleteQuery = " DELETE FROM Carbon_Footprint_Groups WHERE group_ID = ?";
-        db.query(DeleteQuery, [result[0].group_ID], (err, result) => {
-          if(err) {
-            console.log(err)
-            res.status(500).send({ error : 'Something went wrong'})
-          } else {
-            res.status(200).send('Group deleted')
-          }
-        })
+        console.log(result)
+        if (result[0].owner_ID !== req.user.user_id) {
+          res.status(403).send({ error : 'You are not the owner of this group'});
+        } else {
+          const DeleteQuery = " DELETE FROM Carbon_Footprint_Groups WHERE group_ID = ?";
+          db.query(DeleteQuery, [result[0].group_ID], (err, result) => {
+            if(err) {
+              console.log(err)
+              res.status(500).send({ error : 'Something went wrong'})
+            } else {
+              res.status(200).send('Group deleted')
+            }
+          })
+        }
       }
     })
   })
